+++ src/Be.Vlaanderen.Basisregisters.Api/Exceptions/BadRequestResponseExamples.cs
@@ -1,19 +1,19 @@
 namespace Be.Vlaanderen.Basisregisters.Api.Exceptions
 {
-    using Microsoft.AspNetCore.Http;
+    using System.Collections.Generic;
+    using BasicApiProblem;
     using Swashbuckle.AspNetCore.Filters;
-    using ProblemDetails = BasicApiProblem.ProblemDetails;
 
-    public class BadRequestResponseExamples : IExamplesProvider<ProblemDetails>
+    public class BadRequestResponseExamples : IExamplesProvider<ValidationProblemDetails>
     {
-        public ProblemDetails GetExamples() =>
-            new ProblemDetails
+        public ValidationProblemDetails GetExamples() =>
+            new ValidationProblemDetails
             {
-                ProblemTypeUri = "urn:be.vlaanderen.basisregisters.api:bad-request",
-                HttpStatus = StatusCodes.Status400BadRequest,
-                Title = ProblemDetails.DefaultTitle,
-                Detail = "<meer informatie over de foutieve data>",
-                ProblemInstanceUri = ProblemDetails.GetProblemNumber()
+                ValidationErrors = new Dictionary<string, string[]>
+                {
+                    { "Voornaam", new[] {"Veld is verplicht." }},
+                    { "Naam", new[] {"Veld mag niet kleiner zijn dan 4 karakters.", "Veld mag niet groter zijn dan 100 karakters." }}
+                }
             };
     }
 }